+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### Change Log
 
+- [January, 2024] Default model set to `gpt-4-turbo-preview`
+- [January, 2024] OpenAI's January 2024 models supported
 - [January, 2024] Mac/Win Monadic Chat Console UI improved
 - [January, 2024] Image understanging feature
 - [December, 2023] Voice Interpreter app added